Refactor code to separate in modules and move temperature methods to their own file
