+++ src/tempMethods.js
@@ -1,0 +1,16 @@
+const tempDiv = document.getElementById('tempInDegrees');
+
+const convertTempToCelsius = (absoluteTemp) => {
+  tempDiv.textContent = '';
+  const celciusTemp = (absoluteTemp - 273.15).toFixed(2);
+  tempDiv.textContent = `The temperature in degrees Celsius is ${celciusTemp}`;
+};
+
+const convertTempToFahrenheit = (absoluteTemp) => {
+  tempDiv.textContent = '';
+  const celciusTemp = (absoluteTemp - 273.15);
+  const fahrenheitTemp = ((celciusTemp * (9 / 5)) + 32).toFixed(2);
+  tempDiv.textContent = `The temperature in degrees Fahrenheit is ${fahrenheitTemp}`;
+};
+
+export { convertTempToCelsius, convertTempToFahrenheit };
